Fix compile error
We tried getting the function type for a symbol that had a function entry in the environment, but wasn't yet fboundp.
This can happen if the function is in the same compilation unit as the expansion.
Add a quick check to prevent this

Also add a test to catch it. Though this is at compile-time.